Drop support for Go 1.8 and below
