--- server/http_server_go18.go
@@ -1,15 +1,0 @@
-// +build go1.8
-
-package server
-
-import "net/http"
-
-func httpServer(handler http.Handler) *http.Server {
-	return &http.Server{
-		Handler:        handler,
-		MaxHeaderBytes: maxHeaderBytes,
-		ReadTimeout:    readTimeout,
-		WriteTimeout:   writeTimeout,
-		IdleTimeout:    idleTimeout,
-	}
-}
